chore: capitalize constants (#114)
* reproduce

* Fix rounding error test

* Add manual test for reproducing

* Fix test_manual

* I think fix, scale all amounts by MPS

* passes manual and invariant

* Add ValueX7 and MPSLib

* Switch back to uint256

* pre-commit

* Remove old references to AuctionStepLib.MPS

* Clean up

* Scale down after fullMulDiv in _accountPartiallyFilledCheckpoints

* pre-commit

* Rename mul and div to mulUint256 and divUint256

* pre-commit

* Fix order of scaleDown and remove unused parenthesis

* Remove redundant initializations in constructor

* chore: capitalize constants + immutables

* chore: tests

* Change to scaleByMps

* Use totalSupplyX7 internally

* pre-commit

* feedback

* fmt and pre-commit

* Add startCumulativeMps to Bid to simplify

* Simplfiy _calculateFill and add mpsRemainingInAuction to Bid library

* Use the uint256 totalSupply value instead of the scaled up X7 value

* Implement fullMulDiv for ValueX7

* fix

* fix: dirty merge

* chore: doc gen + fmt

---------

+++ test/utils/AuctionBaseTest.sol
@@ -24,7 +24,7 @@ abstract contract AuctionBaseTest is TokenHandler, Test {
     uint256 public constant AUCTION_DURATION = 100;
     uint256 public constant TICK_SPACING = 100;
     uint256 public constant FLOOR_PRICE = 1000 << FixedPoint96.RESOLUTION;
-    uint128 public constant TOTAL_SUPPLY = 1000e18;
+    uint256 public constant TOTAL_SUPPLY = 1000e18;
 
     address public alice;
     address public tokensRecipient;
